Refactors orafin to it's own sub-package
